requirements.txt
fastapi>=0.70.0,<0.111.0
uvicorn[standard]>=0.15.0,<0.25.0
python-multipart>=0.0.5,<0.0.7
opencv-python-headless>=4.5.0,<5.0.0
numpy>=1.18.0,<1.25.0
grpcio>=1.56.0
grpcio-tools>=1.56.0
protobuf>=4.21.0
